Move slogdet to linalg instead of numpy
