+++ keras/src/backend/numpy/numpy.py
@@ -1670,10 +1670,6 @@ def select(condlist, choicelist, default=0):
     return np.select(condlist, choicelist, default=default)
 
 
-def slogdet(x):
-    return tuple(np.linalg.slogdet(x))
-
-
 def argpartition(x, kth, axis=-1):
     return np.argpartition(x, kth, axis).astype("int32")
